Moved analyzer interface
