+++ mango_tango_cib/analyzer_interface/column_automap.py
