+++ mango_tango_cib/analyzer_interface/declaration.py
