+++ mango_tango_cib/analyzers/example/example_base/main.py
@@ -1,6 +1,6 @@
 import polars as pl
 
-from analyzer_interface.context import PrimaryAnalyzerContext
+from mango_tango_cib.analyzer_interface.context import PrimaryAnalyzerContext
 from mango_tango_cib.terminal_tools import ProgressReporter
