+++ mango_tango_cib/analyzers/example/example_report/interface.py
@@ -1,4 +1,8 @@
-from analyzer_interface import AnalyzerOutput, OutputColumn, SecondaryAnalyzerInterface
+from mango_tango_cib.analyzer_interface import (
+    AnalyzerOutput,
+    OutputColumn,
+    SecondaryAnalyzerInterface,
+)
 
 from ..example_base.interface import interface as example_base
