+++ mango_tango_cib/analyzers/hashtags/main.py
@@ -3,7 +3,7 @@ from itertools import accumulate
 
 import polars as pl
 
-from analyzer_interface.context import PrimaryAnalyzerContext
+from mango_tango_cib.analyzer_interface.context import PrimaryAnalyzerContext
 
 from .interface import (
     COL_AUTHOR_ID,
