+++ mango_tango_cib/analyzers/ngram_stats/__init__.py
@@ -1,4 +1,4 @@
-from analyzer_interface import SecondaryAnalyzerDeclaration
+from mango_tango_cib.analyzer_interface import SecondaryAnalyzerDeclaration
 
 from .interface import interface
 from .main import main
