+++ mango_tango_cib/analyzers/ngram_stats/main.py
@@ -2,7 +2,7 @@ import polars as pl
 import pyarrow as pa
 import pyarrow.parquet as pq
 
-from analyzer_interface.context import SecondaryAnalyzerContext
+from mango_tango_cib.analyzer_interface.context import SecondaryAnalyzerContext
 from mango_tango_cib.terminal_tools import ProgressReporter
 
 from ..ngrams.interface import (
