+++ mango_tango_cib/analyzers/ngram_web/__init__.py
@@ -1,4 +1,7 @@
-from analyzer_interface import WebPresenterDeclaration, WebPresenterInterface
+from mango_tango_cib.analyzer_interface import (
+    WebPresenterDeclaration,
+    WebPresenterInterface,
+)
 
 from .factory import factory
 from .interface import interface
