+++ mango_tango_cib/analyzers/ngrams/main.py
@@ -2,7 +2,7 @@ import re
 
 import polars as pl
 
-from analyzer_interface.context import PrimaryAnalyzerContext
+from mango_tango_cib.analyzer_interface.context import PrimaryAnalyzerContext
 from mango_tango_cib.terminal_tools import ProgressReporter
 
 from .interface import (
