+++ mango_tango_cib/analyzers/time_coordination/main.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 
 import polars as pl
 
-from analyzer_interface.context import PrimaryAnalyzerContext
+from mango_tango_cib.analyzer_interface.context import PrimaryAnalyzerContext
 
 from .interface import (
     COL_TIMESTAMP,
