+++ mango_tango_cib/app/analysis_output_context.py
@@ -3,7 +3,7 @@ from typing import Literal, Optional
 
 from pydantic import BaseModel
 
-from analyzer_interface import AnalyzerOutput, SecondaryAnalyzerInterface
+from mango_tango_cib.analyzer_interface import AnalyzerOutput, SecondaryAnalyzerInterface
 from mango_tango_cib.storage import SupportedOutputExtension
 
 from .analysis_context import AnalysisContext
